Modified rubocop to exclude output files Fixed lint issues
