+++ lib/data/books.json
@@ -1,7 +1,12 @@
 {
-  "#<Book:0x000055c10ca6c4f8>": {
+  "#<Book:0x000055bb14e4c378>": {
     "id": 0,
     "title": "wqrt",
     "author": "123qwrf"
+  },
+  "#<Book:0x000055bb14e36438>": {
+    "id": 1,
+    "title": "kljh",
+    "author": "klj"
   }
 }
